# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RS Mas Amsyar WEB API",
        "description": "\r\n * > Footprint 1.0 (2024 June 25) `Developed along with the web frontend to be consumed by and also enabled integration with 3rd party API in the future`\r\n * > Footprint x (2024 xx xx) `this-is-footnote-format-please-continue-on-next-dev`",
        "contact": {
            "name": "Evelline Krist.",
            "email": "ev.attoff@gmail.com"
        },
        "version": "VER 1.0",
        "x-logo": {
            "url": "https://upload.wikimedia.org/wikipedia/commons/thumb/9/9a/Laravel.svg/985px-Laravel.svg.png"
        }
    },
    "paths": {
...
        "/api/satisfaction": {
            "get": {
                "tags": [
                    "Satisfaction"
                ],
                "summary": "Display a listing of items",
                "operationId": "satisfactionIndex",
                "parameters": [
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "current page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    },
                    {
                        "name": "_limit",
                        "in": "query",
                        "description": "max item in a page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 10
                        }
                    },
                    {
                        "name": "_search",
                        "in": "query",
                        "description": "word to search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "_dir",
                        "in": "query",
                        "description": "order by direction",
                        "required": false,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Satisfaction"
                ],
                "summary": "Store a newly created item",
                "operationId": "satisfactionStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "star"
                                ],
                                "properties": {
                                    "star": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Nadjwa Shihab"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "nadjwa172@narasi.co.id"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Nice!"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Saya suka pelayanannya cepat. Saya jalur tunai sih."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            }
        },
        "/api/satisfaction/{id}": {
            "get": {
                "tags": [
                    "Satisfaction"
                ],
                "summary": "Display the specified item",
                "operationId": "satisfactionShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of item that needs to be displayed",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Satisfaction"
                ],
                "summary": "Update the specified item",
                "operationId": "satisfactionUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of item that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "clean-eating"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "star"
                                ],
                                "properties": {
                                    "star": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "[EDITED] Nadjwa Shihab"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "nadjwa172@narasi.co.id"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "[EDITED] Nice!"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "[EDITED] Saya suka pelayanannya cepat. Saya jalur tunai sih."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Satisfaction"
                ],
                "summary": "Remove the specified item",
                "operationId": "satisfactionDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of item that needs to be removed",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "passport_token_ready": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format without \"Bearer \"",
                "name": "Authorization",
                "in": "header"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}